Extract user methods
