+++ RedditSharp/Reddit.cs
@@ -1,10 +1,10 @@
 using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
+using RedditSharp.Models;
+using RedditSharp.Workers;
 using System;
 using System.Linq;
 using System.Net;
-using System.Security.Authentication;
-using RedditSharp.Models;
 using System.Threading.Tasks;
 using DefaultWebAgent = RedditSharp.WebAgent;
 
@@ -16,15 +16,9 @@ namespace RedditSharp
     public class Reddit
     {
         #region Constant Urls
-
-        private const string SslLoginUrl = "https://ssl.reddit.com/api/login";
-        private const string LoginUrl = "/api/login/username";
-        private const string UserInfoUrl = "/user/{0}/about.json";
-        private const string MeUrl = "/api/me.json";
-        private const string OAuthMeUrl = "/api/v1/me.json";
+        
         private const string SubredditAboutUrl = "/r/{0}/about.json";
         private const string ComposeMessageUrl = "/api/compose";
-        private const string RegisterAccountUrl = "/api/register";
         private const string GetModelUrl = "/api/info.json?id={0}";
         private const string GetCommentUrl = "/r/{0}/comments/{1}/foo/{2}";
         private const string GetPostUrl = "{0}.json";
@@ -127,7 +121,8 @@ namespace RedditSharp
         {
             DefaultWebAgent.RootDomain = OAuthDomainUrl;
             WebAgent.AccessToken = accessToken;
-            InitOrUpdateUser();
+            UserWorker userWorker = new UserWorker(this);
+            userWorker.InitOrUpdateUser();
         }
         /// <summary>
         /// Creates a Reddit instance with the given WebAgent implementation
@@ -157,7 +152,11 @@ namespace RedditSharp
                 DefaultValueHandling = DefaultValueHandling.Ignore
             };
             CaptchaSolver = new ConsoleCaptchaSolver();
-            if(initUser) InitOrUpdateUser();
+            if(initUser)
+            {
+                UserWorker userWorker = new UserWorker(this);
+                userWorker.InitOrUpdateUser();
+            }
         }
 
         /// <summary>
@@ -169,67 +168,14 @@ namespace RedditSharp
         /// <returns></returns>
         public AuthenticatedUser LogIn(string username, string password, bool useSsl = true)
         {
-            if (Type.GetType("Mono.Runtime") != null)
-                ServicePointManager.ServerCertificateValidationCallback = (s, c, ch, ssl) => true;
-            WebAgent.Cookies = new CookieContainer();
-            HttpWebRequest request;
-            if (useSsl)
-                request = WebAgent.CreatePost(SslLoginUrl);
-            else
-                request = WebAgent.CreatePost(LoginUrl);
-            var stream = request.GetRequestStream();
-            if (useSsl)
-            {
-                WebAgent.WritePostBody(stream, new
-                {
-                    user = username,
-                    passwd = password,
-                    api_type = "json"
-                });
-            }
-            else
-            {
-                WebAgent.WritePostBody(stream, new
-                {
-                    user = username,
-                    passwd = password,
-                    api_type = "json",
-                    op = "login"
-                });
-            }
-            stream.Close();
-            var response = (HttpWebResponse)request.GetResponse();
-            var result = WebAgent.GetResponseString(response.GetResponseStream());
-            var json = JObject.Parse(result)["json"];
-            if (json["errors"].Count() != 0)
-                throw new AuthenticationException("Incorrect login.");
-
-            InitOrUpdateUser();
-
-            return User;
+            UserWorker userWorker = new UserWorker(this);
+            return userWorker.LogIn(username, password, useSsl);
         }
 
         public RedditUser GetUser(string name)
         {
-            var request = WebAgent.CreateGet(string.Format(UserInfoUrl, name));
-            var response = request.GetResponse();
-            var result = WebAgent.GetResponseString(response.GetResponseStream());
-            var json = JObject.Parse(result);
-            return new RedditUser().Init(this, json, WebAgent);
-        }
-
-        /// <summary>
-        /// Initializes the User property if it's null,
-        /// otherwise replaces the existing user object
-        /// with a new one fetched from reddit servers.
-        /// </summary>
-        public void InitOrUpdateUser()
-        {
-            var request = WebAgent.CreateGet(string.IsNullOrEmpty(WebAgent.AccessToken) ? MeUrl : OAuthMeUrl);
-            var response = (HttpWebResponse)request.GetResponse();
-            var result = WebAgent.GetResponseString(response.GetResponseStream());
-            var json = JObject.Parse(result);
-            User = new AuthenticatedUser().Init(this, json, WebAgent);
+            UserWorker userWorker = new UserWorker(this);
+            return userWorker.GetUser(name);
         }
 
         #region Obsolete Getter Methods
@@ -335,20 +281,8 @@ namespace RedditSharp
         /// <returns>The newly created user account</returns>
         public AuthenticatedUser RegisterAccount(string userName, string passwd, string email = "")
         {
-            var request = WebAgent.CreatePost(RegisterAccountUrl);
-            WebAgent.WritePostBody(request.GetRequestStream(), new
-            {
-                api_type = "json",
-                email = email,
-                passwd = passwd,
-                passwd2 = passwd,
-                user = userName
-            });
-            var response = request.GetResponse();
-            var result = WebAgent.GetResponseString(response.GetResponseStream());
-            var json = JObject.Parse(result);
-            return new AuthenticatedUser().Init(this, json, WebAgent);
-            // TODO: Error
+            UserWorker userWorker = new UserWorker(this);
+            return userWorker.RegisterAccount(userName, passwd, email);
         }
 
         public Model GetModelByFullname(string fullname)
